Fixed sprinting bug and  small update to input system
When holding shift to sprint, oxygen no longer gets decreased when standing still. Also, updated input system so it uses 'value' instead of 'pass through'. for movement and camera. Sprinting is also now 'button' instead of 'pass through'. This is better for when having multiple devices connected to the game (such as keyboard and gamepad at the same time).